+++ app/assets/javascripts/admin/addon/controllers/admin-email-logs-skipped.js
@@ -1,16 +1,9 @@
-import { action } from "@ember/object";
 import { observes } from "@ember-decorators/object";
 import discourseDebounce from "discourse/lib/debounce";
 import { INPUT_DELAY } from "discourse/lib/environment";
 import AdminEmailLogsController from "admin/controllers/admin-email-logs";
 
-export default class AdminEmailBouncedController extends AdminEmailLogsController {
-  @action
-  handleShowIncomingEmail(id, event) {
-    event?.preventDefault();
-    this.send("showIncomingEmail", id);
-  }
-
+export default class AdminEmailLogsSkippedController extends AdminEmailLogsController {
   @observes("filter.{status,user,address,type}")
   filterEmailLogs() {
     discourseDebounce(this, this.loadLogs, INPUT_DELAY);
