--- app/assets/javascripts/admin/addon/routes/admin-email-received.js
@@ -1,5 +1,0 @@
-import AdminEmailIncomings from "admin/routes/admin-email-incomings";
-
-export default class AdminEmailReceivedRoute extends AdminEmailIncomings {
-  status = "received";
-}
