--- app/assets/javascripts/admin/addon/routes/admin-email-sent.js
@@ -1,5 +1,0 @@
-import AdminEmailLogs from "admin/routes/admin-email-logs";
-
-export default class AdminEmailSentRoute extends AdminEmailLogs {
-  status = "sent";
-}
