--- app/assets/javascripts/admin/addon/routes/admin-email-skipped.js
@@ -1,5 +1,0 @@
-import AdminEmailLogs from "admin/routes/admin-email-logs";
-
-export default class AdminEmailSkippedRoute extends AdminEmailLogs {
-  status = "skipped";
-}
