+++ app/assets/javascripts/admin/addon/routes/admin-route-map.js
@@ -28,14 +28,21 @@ export default function () {
     );
 
     this.route(
-      "adminEmail",
-      { path: "/email", resetNamespace: true },
+      "adminEmailLogs",
+      { path: "/email-logs", resetNamespace: true },
       function () {
-        this.route("sent");
+        this.route("sent", { path: "/" });
         this.route("skipped");
         this.route("bounced");
         this.route("received");
         this.route("rejected");
+      }
+    );
+
+    this.route(
+      "adminEmail",
+      { path: "/email", resetNamespace: true },
+      function () {
         this.route("previewDigest", { path: "/preview-digest" });
         this.route("advancedTest", { path: "/advanced-test" });
         this.route(
