+++ app/assets/javascripts/admin/addon/templates/email-logs-bounced.gjs
