+++ app/assets/javascripts/admin/addon/templates/email-logs-received.gjs
