+++ app/assets/javascripts/admin/addon/templates/email-logs-rejected.gjs
