+++ app/assets/javascripts/admin/addon/templates/email-logs-sent.gjs
